pyproject.toml
[project]
name = "nmck"
version = "0.2.1"
description = "A powerful CLI tool to check name availability across domains and platforms"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Nmck Contributors"}
]
keywords = ["cli", "domain", "availability", "github", "gitlab", "pypi", "npm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
    "httpx",
    "click",
    "python-whois",
    "dnspython",
    "rich",
    "python-dotenv",
    "beautifulsoup4>=4.14.3",
]

[project.urls]
Homepage = "https://github.com/cphotor/nmck"
Repository = "https://github.com/cphotor/nmck"
"Bug Tracker" = "https://github.com/cphotor/nmck/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
nmck = "nmck.cli:cli"

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
  "E", "W", "F", "I", "N", "UP", "B", "A", "C4", "T20", "SIM", "ARG", "PTH"
]
fixable = ["ALL"]
ignore = ["E501", "T20"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# VS Code/Windsurf Python environment configuration
[tool.vscode]
python.terminal.activateEnvironment = true
python.terminal.autoActivationType = "shellStartup"
python.workspaceSearchPaths = ["./**/.venv"]
